level2/requirements.txt
annotated-types==0.6.0
pycryptodome==3.19.0
pydantic==2.4.2
pydantic-core==2.10.1
typing-extensions==4.8.0
